Changed builder types to be consuming, and to return objects that implement their traits.
This is going to be one of the more unexpected changes I've made.
The builder types (AsyncDirBuilderTrait and AsyncOpenOptionsTrait)
now are consuming, and are not required to return Self (or &mut self).
Instead, they are required to return new objects that implement their
respective traits.  The reason for this change is because it makes it
possible for a builder type to return a different type of object each
time a new setting is made, if that makes sense in the particular
use case.  I may tweak the design further so that dynamic objects
are returned, but I haven't yet decided if this is a good idea or
not.