Added my own version of the futures_core::stream::Stream trait.
I need this trait in order to make various components of my crate
useful, but at the same time I don't want to bind myself to 'futures',
'futures_lite', 'futures_core', or any other crate.  So, what I've
done is created a new copy of the trait in my own crate, and made it
so that I have blanket implementations of it for every type that
implements one of the external traits.  This decouples my code from
the particular version of the futures crate someone is using, while
making it compatible with all of them.