refactor: I removed my definitions of the async read, write, etc., for futures-io versions.
I didn't realize that the futures crate was actually broken up into
numerous subcrates, including the futures-io crate.  That crate
defines everything that I need for async reading, buffered reading,
writing, and seeking.  It's also used all over the ecosystem, which
means that it's better for me to use it.

The one downside is that I can't create blanket implementations
directly without pulling in the futures-utils crate, which has a
LOT of dependencies that I don't want to deal with.  I'll leave
that as a problem to solve another day.