Added some more global warnings
I will eventually tweak the set of warnings to be far more complete
and conservative, but this was my attempt at getting some behavior
that no longer seems available.  That is, it used to be that when
you ran 'cargo doc' it warned whenever it encountered 'TODO' or
'FIXME'.  This no longer seems to be the case.